backend/services/user-service/inbody_extractor/requirements.txt
opencv-python-headless>=4.8
numpy>=1.24
pydantic>=2.0
Pillow>=10.0
pytesseract>=0.3.13
